Fix flash of unstyled content by reading styles directly from storage
The inject-css content script ran at document_start, but still had to
message the background service worker and wait for it to read
chrome.storage.local before it could insert the <style> tag. When the
MV3 service worker was cold, waking it up added enough latency to
cause a visible flash of unstyled content before styles applied.

The content script now reads chrome.storage.local directly (handled
by the browser process, not the service worker) and applies the
existing URL-matching logic itself, removing the message round trip
and the 300ms x 10 retry-polling fallback that was compensating for it.

Extracted the pure getStylesForPage/BackgroundPageUtils matching logic
into a new shared src/styles package so both the background page and
the content script can use it. Moved badge-icon updates into the
existing tabs.onUpdated listener since they used to piggyback on the
now-removed message, and removed the GetStylesForIframe message type,
which was only ever sent by the old content script.
