+++ src/background/contextmenu.ts
@@ -1,7 +1,6 @@
 import { t } from '@stylebot/i18n';
 import { OpenStylebotFromContextMenu } from '@stylebot/types';
-
-import BackgroundPageUtils from './utils';
+import { BackgroundPageUtils } from '@stylebot/styles';
 
 const CONTEXT_MENU_ID = 'stylebot-contextmenu';
 const VIEW_OPTIONS_MENU_ITEM_ID = 'view-options';
