+++ src/background/styles.ts
@@ -1,7 +1,7 @@
 import * as postcss from 'postcss';
 
 import { getCurrentTimestamp } from '@stylebot/utils';
-import { appendImportantToDeclarations } from '@stylebot/css';
+import { getStylesForPage } from '@stylebot/styles';
 
 import {
   Style,
@@ -10,7 +10,7 @@ import {
   ApplyStylesToTab,
 } from '@stylebot/types';
 
-import BackgroundPageUtils from './utils';
+export { getStylesForPage } from '@stylebot/styles';
 
 export const updateIcon = (
   tab: chrome.tabs.Tab,
@@ -72,51 +72,6 @@ export const getAll = (): Promise<StyleMap> =>
 export const get = async (url: string): Promise<StyleWithoutUrl> => {
   const styles = await getAll();
   return styles[url];
-};
-
-export const getStylesForPage = (
-  pageUrl: string,
-  allStyles: StyleMap,
-  important = false
-): {
-  styles: Array<Style>;
-  defaultStyle?: Style;
-} => {
-  if (!pageUrl) {
-    return { styles: [] };
-  }
-
-  if (!BackgroundPageUtils.isValidHTML(pageUrl)) {
-    return { styles: [] };
-  }
-
-  const styles = [];
-  let defaultStyle: Style | undefined;
-
-  for (const url in allStyles) {
-    const matches = BackgroundPageUtils.matches(pageUrl, url);
-
-    if (matches && allStyles[url]) {
-      const css = important
-        ? appendImportantToDeclarations(allStyles[url].css)
-        : allStyles[url].css;
-
-      const { enabled, readability, modifiedTime } = allStyles[url];
-      const style = { url, css, enabled, readability, modifiedTime };
-
-      if (url !== '*') {
-        if (!defaultStyle || url.length > defaultStyle.url.length) {
-          defaultStyle = style;
-        }
-      }
-
-      if (style.css) {
-        styles.push(style);
-      }
-    }
-  }
-
-  return { styles, defaultStyle };
 };
 
 export const setAll = async (styles: StyleMap): Promise<void> => {
