+++ src/styles/index.ts
@@ -1,0 +1,2 @@
+export { default as BackgroundPageUtils } from './utils';
+export { getStylesForPage } from './page';
